pyproject.toml
[project]
name = "python-sample-junk"
version = "0.0.1"
description = "python sample junk"
authors = [
  {name = "sakito", email = "sakito@sakito.com"},
]
license = "MIT"
readme = "README.md"
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: Unix",
  "Operating System :: POSIX :: Linux",
  "Environment :: Console",
  "Environment :: MacOS X",
  "Framework :: Pytest",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
  "pytest>=8",
  "ruff>=0.8.0",
  "pytz",
  "python-dateutil",
  "python-box",
]

[tool.pytest.ini_options]
testpaths = ["tests",]
pythonpath = [".",]
filterwarnings = ["ignore::DeprecationWarning",]

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]

line-length = 119
indent-width = 4

[tool.ruff.lint]
select = [
  "E4",
  "E7",
  "E9",
  "F", # pyflakes
  "I", # isort
]
ignore = []

fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
